refactor(web): move profile switch to header with sidebar logo animation
- Move user identity (workspace name, email, avatar) from sidebar to header
- Split workspace switch (name + chevron) from user menu (avatar)
- Add Sokosumi branding to sidebar with collapse animation
- Delete old profile-switch components and test
- Create new header profile components:
  - HeaderProfileSection (server): fetches session and user data
  - HeaderProfileSectionClient: wrapper for workspace switch and user menu
  - HeaderWorkspaceSwitch: displays workspace name/email, handles workspace switching
  - HeaderUserMenu: displays avatar, handles account dropdown
- Create SidebarLogo component with animated icon/wordmark based on sidebar state
- Pass session to Sidebar for UserCredits component
- Reuse existing profile-switch behavior, data sources, and translations

Closes SOK-564
