+++ apps/web/src/app/(app)/components/sidebar/index.tsx
@@ -17,7 +17,7 @@ import ChatListsClient from "./components/chat-lists.client";
 import CustomTrigger from "./components/custom-trigger";
 import MenuItems from "./components/menu-items";
 import NewChatTaskActions from "./components/new-chat-task-actions";
-import ProfileSwitch from "./components/profile-switch";
+import SidebarLogo from "./components/sidebar-logo.client";
 
 interface SidebarProps {
   adminMenuEnabled: boolean;
@@ -40,7 +40,7 @@ export default function Sidebar({
     <ShadcnSidebar collapsible="icon">
       <SidebarHeader className="h-[64px] border-b">
         <div className="flex items-center gap-2 pt-1 group-data-[collapsible=icon]:justify-center group-data-[collapsible=icon]:pt-1! group-data-[collapsible=icon]:pl-0!">
-          <ProfileSwitch session={session} />
+          <SidebarLogo />
           <CustomTrigger className="group-data-[collapsible=icon]:hidden" />
         </div>
       </SidebarHeader>
